fix testcases
